rename view to document
