+++ src/app/Api/Core/DAL/Documents/BaseBuilder.cs
@@ -1,0 +1,17 @@
+﻿namespace Api.Core.DAL.Documents
+{
+    public abstract class BaseBuilder<T>
+    {
+        protected T data;
+
+        protected BaseBuilder()
+        {
+
+        }
+
+        public T Build()
+        {
+            return data;
+        }
+    }
+}
