+++ src/app/Api/Core/DAL/Documents/BaseDocument.cs
@@ -1,0 +1,10 @@
+﻿using MongoDB.Bson.Serialization.Attributes;
+
+namespace Api.Core.DAL.Documents
+{
+    public class BaseDocument
+    {
+        [BsonId]
+        public string Id { get; set; }
+    }
+}
