+++ src/app/Api/Core/DAL/Documents/User/UserFactory.cs
@@ -1,0 +1,17 @@
+﻿namespace Api.Core.DAL.Documents.User
+{
+    public class UserFactory
+    {
+        public User UnverifiedUser()
+        {
+            UserBuilder builder = new UserBuilder();
+            return builder.NotVerifiedEmail().Build();
+        }
+
+        public User VerifiedUser()
+        {
+            UserBuilder builder = new UserBuilder();
+            return builder.Build();
+        }
+    }
+}
