--- src/app/Api/Core/DAL/Views/BaseBuilder.cs
@@ -1,17 +1,0 @@
-﻿namespace Api.Core.DAL.Views
-{
-    public abstract class BaseBuilder<T>
-    {
-        protected T data;
-
-        protected BaseBuilder()
-        {
-
-        }
-
-        public T Build()
-        {
-            return data;
-        }
-    }
-}
