--- src/app/Api/Core/DAL/Views/BaseView.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using MongoDB.Bson.Serialization.Attributes;
-
-namespace Api.Core.DAL.Views
-{
-    public class BaseView
-    {
-        [BsonId]
-        public string Id { get; set; }
-        public DateTime CreatedOn { get; set; }
-        public DateTime UpdatedOn { get; set; }
-    }
-}
