--- src/app/Api/Core/DAL/Views/Token/Token.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using Api.Core.Enums;
-
-namespace Api.Core.DAL.Views.Token
-{
-    public class Token : BaseView
-    {
-        public TokenTypeEnum Type { get; set; }
-        public string Value { get; set; }
-        public DateTime ExpireAt { get; set; }
-        public string UserId { get; set; }
-    }
-}
