+++ src/app/Api/Core/Interfaces/DAL/IRepository.cs
@@ -3,21 +3,21 @@ using System.Collections.Generic;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
 using Api.Core.DAL;
-using Api.Core.DAL.Views;
+using Api.Core.DAL.Documents;
 
 namespace Api.Core.Interfaces.DAL
 {
-    public interface IRepository<TView, in TFilter> 
-        where TView : BaseView
+    public interface IRepository<TDocument, in TFilter> 
+        where TDocument : BaseDocument
         where TFilter : BaseFilter
     {
-        Task InsertAsync(TView view);
-        Task InsertManyAsync(IEnumerable<TView> views);
+        Task InsertAsync(TDocument document);
+        Task InsertManyAsync(IEnumerable<TDocument> documents);
 
-        Task<TView> FindOneAsync(TFilter filter);
+        Task<TDocument> FindOneAsync(TFilter filter);
 
-        Task UpdateOneAsync(string id, Expression<Func<TView, object>> fieldSelector, object value);
-        Task UpdateOneAsync(string id, Action<TView> updater);
+        Task UpdateOneAsync(string id, Expression<Func<TDocument, object>> fieldSelector, object value);
+        Task UpdateOneAsync(string id, Action<TDocument> updater);
 
         Task DeleteManyAsync(TFilter filter);
     }
