--- src/app/Api/Core/Interfaces/Services/IViewService.cs
@@ -1,14 +1,0 @@
-﻿using System.Threading.Tasks;
-using Api.Core.DAL;
-using Api.Core.DAL.Views;
-
-namespace Api.Core.Interfaces.Services
-{
-    public interface IViewService<TView, in TFilter>
-        where TView : BaseView
-        where TFilter : BaseFilter
-    {
-        Task<TView> FindByIdAsync(string id);
-        Task<TView> FindOneAsync(TFilter filter);
-    }
-}
