--- src/app/Api/Core/Interfaces/Services/View/IUserService.cs
@@ -1,24 +1,0 @@
-﻿using System.Threading.Tasks;
-using Api.Core.DAL.Repositories;
-using Api.Core.DAL.Views.User;
-using Api.Core.Services.View.Models;
-
-namespace Api.Core.Interfaces.Services.View
-{
-    public interface IUserService : IViewService<User, UserFilter>
-    {
-        Task<User> CreateUserAccountAsync(CreateUserModel model);
-        Task<User> CreateUserAccountAsync(CreateUserGoogleModel model);
-
-        Task<User> FindByEmailAsync(string email);
-
-        Task UpdateLastRequestAsync(string id);
-        Task UpdateResetPasswordTokenAsync(string id, string token);
-        Task UpdatePasswordAsync(string id, string newPassword);
-        Task UpdateInfoAsync(string id, string email, string firstName, string lastName);
-        Task MarkEmailAsVerifiedAsync(string id);
-        Task EnableGoogleAuthAsync(string id);
-
-        Task<bool> IsEmailInUseAsync(string userIdToExclude, string email);
-    }
-}
