+++ src/app/Api/Core/Services/Document/Models/CreateUserGoogleModel.cs
@@ -1,0 +1,9 @@
+﻿namespace Api.Core.Services.Document.Models
+{
+    public class CreateUserGoogleModel
+    {
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+        public string Email { get; set; }
+    }
+}
