+++ src/app/Api/Core/Services/Document/Models/CreateUserModel.cs
@@ -1,0 +1,10 @@
+﻿namespace Api.Core.Services.Document.Models
+{
+    public class CreateUserModel
+    {
+        public string Email { get; set; }
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+        public string Password { get; set; }
+    }
+}
