--- src/app/Api/Core/Services/View/Models/CreateUserGoogleModel.cs
@@ -1,9 +1,0 @@
-﻿namespace Api.Core.Services.View.Models
-{
-    public class CreateUserGoogleModel
-    {
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Email { get; set; }
-    }
-}
