--- src/app/Api/Core/Services/View/Models/CreateUserModel.cs
@@ -1,10 +1,0 @@
-﻿namespace Api.Core.Services.View.Models
-{
-    public class CreateUserModel
-    {
-        public string Email { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Password { get; set; }
-    }
-}
